+++ trunk/NeuralNetwork/src/neuronPackage/NetworkNode.java
@@ -3,7 +3,7 @@ package neuronPackage;
 public interface NetworkNode {
 	void addInput(double val);
 
-	Status advance(double timeStep, int timeofSimulation);
+	Status advance(double timeStep, double timeofSimulation);
 
 	void setCurrentInput();
 }
